create a star outline and star solid component using fontawesome and export to star component
